pin python version to an exact version
this is how it used to be because of the heroku poetry buildpack but i recently switched to using my home computer which has a different default python3. I totally forgot about the buildpack and just widened the range. Anyway now I've added a comment to help my poor memory out for next time.