manually update lockfile
so when there are non-functional changes to the pyproject.toml file (e.g., you remove a comment or move it around or whatever), poetry update does nothing. But the heroku poetry buildpack looks at the lockfile, not at your toml file, so you gotta delete the lockfile and start from scratch OR just manually edit it. i can't believe we're now 4 commits in, this is embarrassing